microservices/notification_service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
redis>=4.5.0
django-redis>=5.2.0
Pillow>=9.0.0
gunicorn>=20.1.0
django-cors-headers>=3.13.0
celery>=5.2.0
sendgrid>=6.9.0
twilio>=7.10.0